pass previewData as a prop to provider
