+++ assets/js/previews/cart.js
@@ -2,19 +2,19 @@
  * External dependencies
  */
 import { __ } from '@wordpress/i18n';
+import { SHIPPING_METHODS_EXIST } from '@woocommerce/block-settings';
 
 /**
  * Internal dependencies
  */
 import productPicture from './product-image';
 import { previewShippingRates } from './shipping-rates';
-
 // Sample data for cart block.
 // This closely resembles the data returned from the Store API /cart endpoint.
 // https://github.com/woocommerce/woocommerce-gutenberg-products-block/tree/master/src/RestApi/StoreApi#cart-api
 export const previewCart = {
 	coupons: [],
-	shipping_rates: previewShippingRates,
+	shipping_rates: SHIPPING_METHODS_EXIST ? previewShippingRates : [],
 	items: [
 		{
 			key: '1',
